+++ mm2src/coins/eth/eth_tests.rs
@@ -227,7 +227,7 @@ fn send_and_refund_erc20_payment() {
 
     let payment = coin
         .send_maker_payment(
-            (now_ms() / 1000) as u32 - 200,
+            now_ms() / 1000 - 200,
             &DEX_FEE_ADDR_RAW_PUBKEY,
             &[1; 20],
             "0.001".parse().unwrap(),
@@ -243,7 +243,7 @@ fn send_and_refund_erc20_payment() {
     let refund = coin
         .send_maker_refunds_payment(
             &payment.tx_hex(),
-            (now_ms() / 1000) as u32 - 200,
+            now_ms() / 1000 - 200,
             &DEX_FEE_ADDR_RAW_PUBKEY,
             &[1; 20],
             &coin.swap_contract_address(),
@@ -288,7 +288,7 @@ fn send_and_refund_eth_payment() {
 
     let payment = coin
         .send_maker_payment(
-            (now_ms() / 1000) as u32 - 200,
+            now_ms() / 1000 - 200,
             &DEX_FEE_ADDR_RAW_PUBKEY,
             &[1; 20],
             "0.001".parse().unwrap(),
@@ -304,7 +304,7 @@ fn send_and_refund_eth_payment() {
     let refund = coin
         .send_maker_refunds_payment(
             &payment.tx_hex(),
-            (now_ms() / 1000) as u32 - 200,
+            now_ms() / 1000 - 200,
             &DEX_FEE_ADDR_RAW_PUBKEY,
             &[1; 20],
             &coin.swap_contract_address(),
